Add dazzling visual enhancements
